#15 refact : html 형식 변환을 훅으로 분리
