+++ src/components/pages/CategoryPage.js
@@ -14,6 +14,7 @@ import btnClose from "../../images/btn_close.svg";
 import icoSearchGray from "../../images/ico_search_gray.svg";
 import icoLogout from "../../images/ico_logout.svg";
 import btnTop from "../../images/btn_top.svg";
+import useCleanHTML from "../../hooks/useCleanHtml";
 
 function CategoryPage() {
   const { categoryId } = useParams();
@@ -27,6 +28,9 @@ function CategoryPage() {
   const pageSize = 12;
 
   const [setCategories] = useState([]);
+
+  // HTML 클린업 훅 사용
+  const { cleanHTMLContent } = useCleanHTML();
 
   useEffect(() => {
     axios
@@ -60,47 +64,6 @@ function CategoryPage() {
       });
   }, [categoryId, page, pageSize]);
 
-  const stripImagesAndFigcaptionAndFigure = (htmlContent) => {
-    let contentWithoutImages = htmlContent.replace(/<img[^>]*>/g, "");
-    let contentWithoutFigcaption = contentWithoutImages.replace(
-      /<figcaption[^>]*>(.*?)<\/figcaption>/gi,
-      ""
-    );
-
-    return contentWithoutFigcaption.replace(/<figure[^>]*>\s*<\/figure>/gi, "");
-  };
-
-  const stripTags = (htmlContent, tagToRemove) => {
-    const tagPattern = new RegExp(
-      `<${tagToRemove}[^>]*>(.*?)<\/${tagToRemove}>`,
-      "gi"
-    );
-    return htmlContent.replace(tagPattern, "");
-  };
-
-  const parseTableData = (htmlContent) => {
-    const parser = new DOMParser();
-    const doc = parser.parseFromString(htmlContent, "text/html");
-    const table = doc.querySelector("table");
-
-    if (!table) return htmlContent;
-
-    let formattedData = "";
-
-    const rows = table.querySelectorAll("tr");
-    rows.forEach((row, rowIndex) => {
-      const cells = row.querySelectorAll("td");
-      if (cells.length === 6 && rowIndex > 0) {
-        formattedData += `
-          주요도시: ${cells[0].innerText.trim()} - 기온: ${cells[1].innerText.trim()} - 날씨: ${cells[2].innerText.trim()}<br/>
-          주요도시: ${cells[3].innerText.trim()} - 기온: ${cells[4].innerText.trim()} - 날씨: ${cells[5].innerText.trim()}<br/><br/>
-        `;
-      }
-    });
-
-    return formattedData;
-  };
-
   const handlePageChange = (newPage) => {
     if (newPage > 0 && newPage <= totalPages) {
       setPage(newPage);
@@ -127,17 +90,7 @@ function CategoryPage() {
                       <li className="txt">
                         <div
                           dangerouslySetInnerHTML={{
-                            __html: (() => {
-                              const contentWithoutImagesAndFigcaptionAndFigure =
-                                stripImagesAndFigcaptionAndFigure(news.content);
-                              const contentWithoutH2 = stripTags(
-                                contentWithoutImagesAndFigcaptionAndFigure,
-                                "h2"
-                              );
-                              const formattedTableData =
-                                parseTableData(contentWithoutH2);
-                              return formattedTableData;
-                            })(),
+                            __html: cleanHTMLContent(news.content), // 커스텀 훅 사용
                           }}
                         />
                       </li>
